Rename pipeapi package to pipegql
The package holds the pipe's GraphQL field selection, response DTOs, and
unit conversions, not a client. Naming it pipegql keeps it distinct from
the transport client in client/api_client.go and from the tfsdk model and
schema, which both already carry those names in the resource.
